feat: add mapping types for scheduled task DTO and update trigger handling
