+++ apps/chat/src/utils/scheduled-task-trigger.ts
@@ -9,6 +9,10 @@ import type {
   ScheduleTriggerDto,
   UpdateScheduledTaskBodyDto,
 } from '@epam/chat-api-client';
+import {
+  UnsupportedTriggerReason,
+  type ScheduledTaskDtoMappingResult,
+} from '../types/scheduled-task-mapping';
 import { apSchedulerDayToJsDay, jsDayToApSchedulerDay } from './cron-weekday';
 
 /** Which boundary of a cron activity window is being computed. */
@@ -16,21 +20,6 @@ enum CronWindowEdge {
   Start = 'start',
   End = 'end',
 }
-
-/** Why a {@link ScheduledTaskDto} could not be mapped back to editable form values. */
-export enum UnsupportedTriggerReason {
-  /** `trigger.cron.fields` uses a shape (extra keys, or both `day_of_week` and `day`) the editor's schedule controls cannot represent. */
-  UnsupportedCronShape = 'unsupportedCronShape',
-  /** Neither or both of `trigger.date`/`trigger.cron` are set, so the schedule type cannot be determined. */
-  UnsupportedTriggerType = 'unsupportedTriggerType',
-  /** The task is missing `model` or `prompt`, so it cannot be represented as a valid update payload. */
-  MissingRequiredFields = 'missingRequiredFields',
-}
-
-/** Result of mapping a {@link ScheduledTaskDto} back to editable form values. */
-export type ScheduledTaskDtoMappingResult =
-  | { ok: true; values: ScheduledTaskCreateFormValues }
-  | { ok: false; reason: UnsupportedTriggerReason };
 
 const CRON_FIELD_KEYS = new Set(['hour', 'minute', 'day_of_week', 'day']);
